Refactor: Improve SDK package metadata
Deleted the existing readme file b/README.md and the contents of file src/Directory.Build.targets. Changed the company name in the metadata to DGMJR-IO, changed the product name to DGMJR-IO Core Libraries, and updated the slug for the readme to include the project name. Added new settings to define package metadata, including keywords, categories, and tags. Updated the UsingsPackageReference to include additional properties.